raspberry_pi_display/requirements.txt
requests>=2.25.0
pygame>=2.0.0
python-dotenv>=0.19.0